[REFACTOR] entity datatype 변경

+++ src/main/kotlin/com/alloon/alloonserver/domain/feed/Feed.kt
@@ -4,7 +4,6 @@ import com.alloon.alloonserver.domain.base.BasePermanentEntity
 import com.alloon.alloonserver.domain.mission.Mission
 import com.alloon.alloonserver.domain.mission.MissionMember
 import jakarta.persistence.*
-import java.math.BigInteger
 
 @Entity
 class Feed(
@@ -12,7 +11,7 @@ class Feed(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "feed_id")
-    val id: BigInteger? = null,
+    val id: Long? = null,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "mission_member_id", nullable = false)
