+++ src/main/kotlin/com/alloon/alloonserver/domain/mission/Hashtag.kt
@@ -1,13 +1,7 @@
 package com.alloon.alloonserver.domain.mission
 
 import com.alloon.alloonserver.domain.base.BasePermanentEntity
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-import java.math.BigInteger
-import java.time.LocalDateTime
+import jakarta.persistence.*
 
 @Entity
 class Hashtag(
@@ -15,7 +9,7 @@ class Hashtag(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "hashtag_id")
-    val id: BigInteger? = null,
+    val id: Long? = null,
 
     @Column(nullable = false, length = 31)
     val hashtag: String,
