+++ src/main/kotlin/com/alloon/alloonserver/domain/report/Block.kt
@@ -3,7 +3,6 @@ package com.alloon.alloonserver.domain.report
 import com.alloon.alloonserver.domain.base.BasePermanentEntity
 import com.alloon.alloonserver.domain.user.User
 import jakarta.persistence.*
-import java.math.BigInteger
 
 @Entity
 class Block(
@@ -11,7 +10,7 @@ class Block(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "block_id")
-    val id: BigInteger? = null,
+    val id: Long? = null,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "blocker_id", nullable = false)
