+++ src/main/kotlin/com/alloon/alloonserver/domain/user/Contact.kt
@@ -2,8 +2,6 @@ package com.alloon.alloonserver.domain.user
 
 import com.alloon.alloonserver.domain.base.BaseEntity
 import jakarta.persistence.*
-import java.math.BigInteger
-import java.time.LocalDateTime
 
 @Entity
 class Contact(
@@ -11,7 +9,7 @@ class Contact(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "contact_id")
-    val id: BigInteger? = null,
+    val id: Long? = null,
 
     @Column(nullable = false, unique = true, length = 127)
     val email: String,
